+++ includes/minishell.h
@@ -4,7 +4,7 @@
 # include "libft/libft.h"
 # include <readline/readline.h>
 # include <readline/history.h>
-# include "get_next_line.h"
+# include "get_next_line/get_next_line.h"
 # include <unistd.h>
 # include <fcntl.h>
 # include <dirent.h>
@@ -93,15 +93,15 @@ t_tree_node *fill_tree(t_token *start, t_token *end);
 t_token *divide_tokens(t_token *start, t_token *end, t_token **left, t_token **right);
 t_token *find_lowest_priority(t_token *start, t_token *end);
 t_tree_node *create_tree_node(t_token *token);
-void init_prior(t_priora *prior);
+void init_token_priority(t_priora *prior);
 int get_token_priority(t_token_type type, t_priora priority_map);
 t_token *analyze_parenthesis(t_token *tokens, int parenth_open);
-t_token *check_closed_parenths(t_token *last, int parenth_open);
 t_token *find_list_end(t_token *start);
-t_node_type define_node_type(t_token_type tt);
-t_token *subshell_trim(t_token *start, t_token *end, t_token **left);
+//t_node_type define_node_type(t_token_type tt);
+//t_token *subshell_trim(t_token *start, t_token *end, t_token **left);
 void check_if_word_sequence(t_token **priora, t_token **priora_end);
-void assign_value_to_argv(t_tree_node *node, t_token *token);
+//void assign_value_to_argv(t_tree_node *node, t_token *token);
+t_tree_node *parser(char *input);
 
 //test
 void    test_list_data(t_node *test_node);
