+++ main.c
@@ -4,8 +4,8 @@
 int	main(void)
 {
 	char *input;
-	t_token *tokens;
-	t_tree_node *tree;
+	// t_token *tokens;
+	t_tree_node *tree; 
 
 	while (1)
 	{
@@ -13,12 +13,14 @@ int	main(void)
 		add_history((const char*)input);
 		if (!input)
 			continue ;
-		tokens = lexer(input);
+		tree = parser(input);
+/* 		tokens = lexer(input);
 		test_print_tokens(tokens);
 		if (!analyze_parenthesis(tokens, 0))
 			printf("Error\n");
 		tree = fill_tree(tokens, NULL);
-		draw_tree(tree);
+		draw_tree(tree); */
+		//draw_tree(tree);
 	}
 	return (0);
 }
